requirements.txt
streamlit
pandas
fpdf
requests
Pillow
matplotlib
plotly
numpy
openpyxl
LanusStats
kaleido
python-dotenv